+++ src/components/Schematics/Preview-Dynamic.vue
@@ -25,7 +25,12 @@ export default {
   },
   computed: {
     resultExists() {
-      return this.task.serialized_state && this.task.serialized_state._result
+      return this.task?.serialized_state?._result
+    },
+    expectedRuns() {
+      return (
+        this.task?.serialized_state?.n_map_states?.toLocaleString() || 'Unknown'
+      )
     }
   },
   methods: {
@@ -39,6 +44,36 @@ export default {
             } !important`
           : ''
       }
+    },
+    stateClass(state) {
+      const lightStates = [
+        'Submitted',
+        'Cancelled',
+        'Cancelling',
+        'Queued',
+        'Pending'
+      ]
+
+      const textColor = lightStates.includes(state)
+        ? ['grey--text', 'text--darken-4']
+        : ['white--text']
+
+      return [state, ...textColor]
+    }
+  },
+  apollo: {
+    mappedChildren: {
+      query: require('@/graphql/MappedTasks/mapped-children.gql'),
+      variables() {
+        return {
+          taskRunId: this.task.id
+        }
+      },
+      skip() {
+        return this.task.state !== 'Mapped' || !this.task.id
+      },
+      pollInterval: 3000,
+      update: data => data.mapped_children
     }
   }
 }
@@ -57,8 +92,10 @@ export default {
           Task Run
         </v-list-item-subtitle>
         <v-list-item-title>
-          {{ task.flow_run_name }} - {{ task.task.name }}
-          <span v-if="task.map_index > -1"> ({{ task.map_index }})</span>
+          <span>
+            {{ task.flow_run_name }} -
+            {{ task.name ? task.name : task.task.name }}
+          </span>
         </v-list-item-title>
         <v-list-item-subtitle class="caption">
           <v-tooltip top>
@@ -139,10 +176,10 @@ export default {
           <span class="black--text">Expected Runs:</span>
         </v-col>
         <v-col cols="6" class="text-right pt-0">
-          <v-tooltip top>
+          <v-tooltip max-width="300px" top>
             <template #activator="{ on }">
               <span v-on="on">
-                {{ task.serialized_state.n_map_states || 'Unknown' }}
+                {{ expectedRuns }}
               </span>
             </template>
             <span v-if="!task.serialized_state.n_map_states">
@@ -152,7 +189,7 @@ export default {
             <span v-else>
               The number of mapped children expected to run. Note that the
               number of active mapped runs may be less than this if some have
-              not begun running yet.
+              not yet entered a <code>Pending</code> state.
             </span>
           </v-tooltip>
         </v-col>
@@ -188,6 +225,28 @@ export default {
         </v-col>
       </v-row>
     </v-card-text>
+
+    <v-card-actions v-if="mappedChildren" class="pl-3">
+      <div>
+        <div class="body-1 black--text">Mapped Runs</div>
+
+        <v-chip-group column>
+          <v-chip
+            v-for="state in Object.keys(mappedChildren.state_counts)"
+            :key="state"
+            class="px-4 font-weight-bold"
+            :class="stateClass(state)"
+            label
+            small
+          >
+            {{ state }}
+            <span class="font-weight-medium ml-1">
+              ({{ mappedChildren.state_counts[state].toLocaleString() }})
+            </span>
+          </v-chip>
+        </v-chip-group>
+      </div>
+    </v-card-actions>
 
     <v-card-actions v-if="runs && runs.length > 0" class="px-0 py-0">
       <v-list dense style="width: 100%;" class="py-0">