+++ src/pages/TaskRun/Dependencies-Tile.vue
@@ -3,6 +3,57 @@ import CardTitle from '@/components/Card-Title'
 import { mapGetters } from 'vuex'
 import SchematicFlow from '@/components/Schematics/Schematic-Flow'
 
+import gql from 'graphql-tag'
+
+const taskRunFields = `
+  id
+  task_run_id: id
+  map_index
+  name
+  state
+  task_id
+  serialized_state
+  state_message
+  state_timestamp
+  start_time
+  end_time
+  map_index
+
+  task {
+    id
+    mapped
+    name
+    max_retries
+    retry_delay
+    trigger
+    type
+
+    upstream_edges {
+      upstream_task {
+        id
+        name
+      }
+
+      downstream_task {
+        id
+        name
+      }
+    }
+
+    downstream_edges {
+      upstream_task {
+        id
+        name
+      }
+
+      downstream_task {
+        id
+        name
+      }
+    }
+  }
+`
+
 export default {
   filters: {
     typeClass: val => val.split('.').pop()
@@ -12,28 +63,14 @@ export default {
     SchematicFlow
   },
   props: {
-    downstreamCount: {
-      type: Number,
-      required: false,
-      default: () => null
-    },
-    flowRunId: {
-      type: String,
-      required: true
-    },
     loading: {
       type: Boolean,
       required: false,
       default: () => false
     },
-    taskIds: {
-      type: Array,
+    taskRun: {
+      type: Object,
       required: true
-    },
-    upstreamCount: {
-      type: Number,
-      required: false,
-      default: () => null
     }
   },
   data() {
@@ -47,50 +84,106 @@ export default {
   computed: {
     ...mapGetters('user', ['timezone']),
     subtitle() {
-      return `${this.upstreamCount} Upstream • ${this.downstreamCount} Downstream`
+      return `${this.taskRun?.task?.upstream_edges?.length} Upstream • ${this.taskRun?.task?.downstream_edges?.length} Downstream`
+    },
+    query() {
+      if (!this.taskRun) return
+      let queryString = ''
+
+      if (this.taskRun.task.upstream_edges.length) {
+        const upstream = this.taskRun.task.upstream_edges.reduce(
+          (string, edge, i) => {
+            string += `
+          upstream_task_run_${i}: task_run(where: { flow_run_id: { _eq: "${this.taskRun?.flow_run?.id}" }, task_id: { _eq: "${edge?.upstream_task?.id}" }, map_index: { _in: [ -1, ${this.taskRun?.map_index} ] } }) {
+            ${taskRunFields}
+          }
+        `
+            return string
+          },
+          ''
+        )
+        queryString += upstream
+      }
+
+      if (this.taskRun.task.downstream_edges.length) {
+        const downstream = this.taskRun.task.downstream_edges.reduce(
+          (string, edge, i) => {
+            string += `
+          downstream_task_run_${i}: task_run(where: { flow_run_id: { _eq: "${this.taskRun?.flow_run?.id}" }, task_id: { _eq: "${edge?.downstream_task?.id}" }, map_index: { _in: [ -1, ${this.taskRun?.map_index} ] } }) {
+            ${taskRunFields}
+          }
+        `
+            return string
+          },
+          ''
+        )
+        queryString += downstream
+      }
+
+      return queryString == ''
+        ? null
+        : gql`
+        query TaskRunDependencies {
+          ${queryString}
+        }
+      `
+    },
+    modifiedTaskRun() {
+      return {
+        ...this.taskRun,
+        flow_run_name: this.taskRun.flow_run.name,
+        task: {
+          ...this.taskRun.task,
+          task_run_id: this.taskRun.id,
+          name:
+            this.taskRun.name ??
+            this.taskRun.task.name +
+              (this.taskRun.map_index > -1
+                ? ` (${this.taskRun.map_index})`
+                : '')
+        }
+      }
     }
   },
   watch: {},
   methods: {},
   apollo: {
-    flowRun: {
-      query: require('@/graphql/Schematics/flow-run.gql'),
-      variables() {
-        return {
-          id: this.flowRunId
-        }
+    taskRunDependencies: {
+      query() {
+        return this.query
       },
       skip() {
         return (
-          !this.flowRunId ||
-          (this.tasks?.length && this.flowRun.state !== 'Running')
-          // Had to add the second portion here because this
-          // query was running out of control for some reason
+          !this.taskRun?.task?.downstream_edges?.length &&
+          !this.taskRun?.task?.upstream_edges?.length
         )
       },
       update(data) {
-        if (data.flow_run && data.flow_run.length) {
-          let taskRuns = data.flow_run[0].task_runs.filter(tr =>
-            this.taskIds.includes(tr.task.id)
+        const dependencies = Object.keys(data).map(key => {
+          const task = data[key][0]
+
+          // We filter these because the references don't exist on this subset of tasks
+          task.task.upstream_edges = task.task.upstream_edges.filter(
+            edge => this.taskRun.task.id == edge.upstream_task.id
           )
 
-          taskRuns.forEach(tr => {
-            tr.flow_run_name = data.flow_run[0].name
-            tr.task.upstream_edges = tr.task.upstream_edges.filter(edge =>
-              this.taskIds.includes(edge.upstream_task.id)
-            )
-            tr.task.downstream_edges = tr.task.downstream_edges.filter(edge =>
-              this.taskIds.includes(edge.downstream_task.id)
-            )
-          })
+          task.task.downstream_edges = task.task.downstream_edges.filter(
+            edge => this.taskRun.task.id == edge.downstream_task.id
+          )
+          return task
+        })
 
-          this.tasks = taskRuns
-          return data.flow_run[0]
-        } else {
-          this.tasks = []
-        }
+        dependencies.forEach(d => {
+          d.flow_run_name = this.taskRun.flow_run.name
+
+          d.task.name =
+            d.name ?? d.task.name + (d.map_index > -1 ? `(${d.map_index})` : '')
+        })
+        this.tasks = [this.modifiedTaskRun, ...dependencies]
+
+        return data
       },
-      pollInterval: 3000
+      pollInterval: 5000
     }
   }
 }