+++ src/router.js
@@ -32,15 +32,6 @@ export const routes = [
     beforeEnter: multiguard([authNavGuard, tenantNavGuard])
   },
   {
-    name: 'accept',
-    path: '/accept',
-    component: () =>
-      import(
-        /* webpackChunkName: "accept" */ '@/pages/AcceptInvitationPage.vue'
-      ),
-    beforeEnter: multiguard([authNavGuard, tenantNavGuard])
-  },
-  {
     name: 'project',
     path: '/:tenant?/project/:id',
     component: () =>
@@ -235,8 +226,21 @@ export const routes = [
           import(
             /* webpackChunkName: "onboard--resources" */ '@/pages/Onboard/Resources.vue'
           )
+      },
+      {
+        name: 'accept',
+        path: 'accept',
+        component: () =>
+          import(
+            /* webpackChunkName: "accept" */ '@/pages/Onboard/AcceptInvitationPage.vue'
+          ),
+        beforeEnter: multiguard([authNavGuard, tenantNavGuard])
       }
     ]
+  },
+  {
+    path: '/accept',
+    redirect: 'welcome/accept'
   },
   // --------------------------- //
   //